Fixing requested changes
