+++ src/pages/main/Main.tsx
@@ -1,22 +1,17 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import Ticker from 'react-ticker';
 import 'config/i18n';
 
 import Button from 'components/button/Button';
 import routerPaths from 'components/navigation/paths';
-import tickerVars from 'variables/ticker';
+import MemoTicker from 'components/ticker/MemoTicker';
 
 import style from './Main.module.scss';
 
 function MainPage(): JSX.Element {
   const { t } = useTranslation();
   const navigate = useNavigate();
-
-  const tickerPhrases: string[] = t('ticker', { returnObjects: true });
-
-  const MemoTicker = React.memo(Ticker);
 
   return (
     <div className={style.content}>
@@ -38,15 +33,7 @@ function MainPage(): JSX.Element {
         />
       </div>
       <div className={style.bottom}>
-        {tickerPhrases.map((el, index) => (
-          <MemoTicker
-            offset={(index + 1) * tickerVars.baseOffset}
-            speed={tickerVars.speed}
-            mode="await"
-          >
-            {(): JSX.Element => <p>{el}</p>}
-          </MemoTicker>
-        ))}
+        <MemoTicker />
       </div>
     </div>
   );
